requirements.txt
requests
git+https://github.com/Rapptz/discord-ext-menus.git
discord-ui
git+https://github.com/Pycord-Development/pycord
